Split a data frame into a named list by grouping columns

Groups a data frame and splits it, automatically naming each element of the resulting list from the unique group key values. Unlike the bare dplyr::group_split(), the returned list is always named.

Usage

split_by(data, ..., .sep = "_")

Arguments

data

A data frame or tibble.

...

Grouping columns (unquoted, tidy-select).

.sep

Separator used to concatenate multiple group key values into a single list name. Default: "_".

Value

A named list of tibbles.

See also

Examples

split_by(ggplot2::mpg, manufacturer)
#> <list_of<
#>   tbl_df<
#>     manufacturer: character
#>     model       : character
#>     displ       : double
#>     year        : integer
#>     cyl         : integer
#>     trans       : character
#>     drv         : character
#>     cty         : integer
#>     hwy         : integer
#>     fl          : character
#>     class       : character
#>   >
#> >[15]>
#> $audi
#> # A tibble: 18 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 audi         a4           1.8  1999     4 auto… f        18    29 p     comp…
#>  2 audi         a4           1.8  1999     4 manu… f        21    29 p     comp…
#>  3 audi         a4           2    2008     4 manu… f        20    31 p     comp…
#>  4 audi         a4           2    2008     4 auto… f        21    30 p     comp…
#>  5 audi         a4           2.8  1999     6 auto… f        16    26 p     comp…
#>  6 audi         a4           2.8  1999     6 manu… f        18    26 p     comp…
#>  7 audi         a4           3.1  2008     6 auto… f        18    27 p     comp…
#>  8 audi         a4 quattro   1.8  1999     4 manu… 4        18    26 p     comp…
#>  9 audi         a4 quattro   1.8  1999     4 auto… 4        16    25 p     comp…
#> 10 audi         a4 quattro   2    2008     4 manu… 4        20    28 p     comp…
#> 11 audi         a4 quattro   2    2008     4 auto… 4        19    27 p     comp…
#> 12 audi         a4 quattro   2.8  1999     6 auto… 4        15    25 p     comp…
#> 13 audi         a4 quattro   2.8  1999     6 manu… 4        17    25 p     comp…
#> 14 audi         a4 quattro   3.1  2008     6 auto… 4        17    25 p     comp…
#> 15 audi         a4 quattro   3.1  2008     6 manu… 4        15    25 p     comp…
#> 16 audi         a6 quattro   2.8  1999     6 auto… 4        15    24 p     mids…
#> 17 audi         a6 quattro   3.1  2008     6 auto… 4        17    25 p     mids…
#> 18 audi         a6 quattro   4.2  2008     8 auto… 4        16    23 p     mids…
#> 
#> $chevrolet
#> # A tibble: 19 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 chevrolet    c1500 sub…   5.3  2008     8 auto… r        14    20 r     suv  
#>  2 chevrolet    c1500 sub…   5.3  2008     8 auto… r        11    15 e     suv  
#>  3 chevrolet    c1500 sub…   5.3  2008     8 auto… r        14    20 r     suv  
#>  4 chevrolet    c1500 sub…   5.7  1999     8 auto… r        13    17 r     suv  
#>  5 chevrolet    c1500 sub…   6    2008     8 auto… r        12    17 r     suv  
#>  6 chevrolet    corvette     5.7  1999     8 manu… r        16    26 p     2sea…
#>  7 chevrolet    corvette     5.7  1999     8 auto… r        15    23 p     2sea…
#>  8 chevrolet    corvette     6.2  2008     8 manu… r        16    26 p     2sea…
#>  9 chevrolet    corvette     6.2  2008     8 auto… r        15    25 p     2sea…
#> 10 chevrolet    corvette     7    2008     8 manu… r        15    24 p     2sea…
#> 11 chevrolet    k1500 tah…   5.3  2008     8 auto… 4        14    19 r     suv  
#> 12 chevrolet    k1500 tah…   5.3  2008     8 auto… 4        11    14 e     suv  
#> 13 chevrolet    k1500 tah…   5.7  1999     8 auto… 4        11    15 r     suv  
#> 14 chevrolet    k1500 tah…   6.5  1999     8 auto… 4        14    17 d     suv  
#> 15 chevrolet    malibu       2.4  1999     4 auto… f        19    27 r     mids…
#> 16 chevrolet    malibu       2.4  2008     4 auto… f        22    30 r     mids…
#> 17 chevrolet    malibu       3.1  1999     6 auto… f        18    26 r     mids…
#> 18 chevrolet    malibu       3.5  2008     6 auto… f        18    29 r     mids…
#> 19 chevrolet    malibu       3.6  2008     6 auto… f        17    26 r     mids…
#> 
#> $dodge
#> # A tibble: 37 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 dodge        caravan 2…   2.4  1999     4 auto… f        18    24 r     mini…
#>  2 dodge        caravan 2…   3    1999     6 auto… f        17    24 r     mini…
#>  3 dodge        caravan 2…   3.3  1999     6 auto… f        16    22 r     mini…
#>  4 dodge        caravan 2…   3.3  1999     6 auto… f        16    22 r     mini…
#>  5 dodge        caravan 2…   3.3  2008     6 auto… f        17    24 r     mini…
#>  6 dodge        caravan 2…   3.3  2008     6 auto… f        17    24 r     mini…
#>  7 dodge        caravan 2…   3.3  2008     6 auto… f        11    17 e     mini…
#>  8 dodge        caravan 2…   3.8  1999     6 auto… f        15    22 r     mini…
#>  9 dodge        caravan 2…   3.8  1999     6 auto… f        15    21 r     mini…
#> 10 dodge        caravan 2…   3.8  2008     6 auto… f        16    23 r     mini…
#> # ℹ 27 more rows
#> 
#> $ford
#> # A tibble: 25 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 ford         expeditio…   4.6  1999     8 auto… r        11    17 r     suv  
#>  2 ford         expeditio…   5.4  1999     8 auto… r        11    17 r     suv  
#>  3 ford         expeditio…   5.4  2008     8 auto… r        12    18 r     suv  
#>  4 ford         explorer …   4    1999     6 auto… 4        14    17 r     suv  
#>  5 ford         explorer …   4    1999     6 manu… 4        15    19 r     suv  
#>  6 ford         explorer …   4    1999     6 auto… 4        14    17 r     suv  
#>  7 ford         explorer …   4    2008     6 auto… 4        13    19 r     suv  
#>  8 ford         explorer …   4.6  2008     8 auto… 4        13    19 r     suv  
#>  9 ford         explorer …   5    1999     8 auto… 4        13    17 r     suv  
#> 10 ford         f150 pick…   4.2  1999     6 auto… 4        14    17 r     pick…
#> # ℹ 15 more rows
#> 
#> $honda
#> # A tibble: 9 × 11
#>   manufacturer model displ  year   cyl trans      drv     cty   hwy fl    class 
#>   <chr>        <chr> <dbl> <int> <int> <chr>      <chr> <int> <int> <chr> <chr> 
#> 1 honda        civic   1.6  1999     4 manual(m5) f        28    33 r     subco…
#> 2 honda        civic   1.6  1999     4 auto(l4)   f        24    32 r     subco…
#> 3 honda        civic   1.6  1999     4 manual(m5) f        25    32 r     subco…
#> 4 honda        civic   1.6  1999     4 manual(m5) f        23    29 p     subco…
#> 5 honda        civic   1.6  1999     4 auto(l4)   f        24    32 r     subco…
#> 6 honda        civic   1.8  2008     4 manual(m5) f        26    34 r     subco…
#> 7 honda        civic   1.8  2008     4 auto(l5)   f        25    36 r     subco…
#> 8 honda        civic   1.8  2008     4 auto(l5)   f        24    36 c     subco…
#> 9 honda        civic   2    2008     4 manual(m6) f        21    29 p     subco…
#> 
#> $hyundai
#> # A tibble: 14 × 11
#>    manufacturer model   displ  year   cyl trans    drv     cty   hwy fl    class
#>    <chr>        <chr>   <dbl> <int> <int> <chr>    <chr> <int> <int> <chr> <chr>
#>  1 hyundai      sonata    2.4  1999     4 auto(l4) f        18    26 r     mids…
#>  2 hyundai      sonata    2.4  1999     4 manual(… f        18    27 r     mids…
#>  3 hyundai      sonata    2.4  2008     4 auto(l4) f        21    30 r     mids…
#>  4 hyundai      sonata    2.4  2008     4 manual(… f        21    31 r     mids…
#>  5 hyundai      sonata    2.5  1999     6 auto(l4) f        18    26 r     mids…
#>  6 hyundai      sonata    2.5  1999     6 manual(… f        18    26 r     mids…
#>  7 hyundai      sonata    3.3  2008     6 auto(l5) f        19    28 r     mids…
#>  8 hyundai      tiburon   2    1999     4 auto(l4) f        19    26 r     subc…
#>  9 hyundai      tiburon   2    1999     4 manual(… f        19    29 r     subc…
#> 10 hyundai      tiburon   2    2008     4 manual(… f        20    28 r     subc…
#> 11 hyundai      tiburon   2    2008     4 auto(l4) f        20    27 r     subc…
#> 12 hyundai      tiburon   2.7  2008     6 auto(l4) f        17    24 r     subc…
#> 13 hyundai      tiburon   2.7  2008     6 manual(… f        16    24 r     subc…
#> 14 hyundai      tiburon   2.7  2008     6 manual(… f        17    24 r     subc…
#> 
#> $jeep
#> # A tibble: 8 × 11
#>   manufacturer model       displ  year   cyl trans drv     cty   hwy fl    class
#>   <chr>        <chr>       <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#> 1 jeep         grand cher…   3    2008     6 auto… 4        17    22 d     suv  
#> 2 jeep         grand cher…   3.7  2008     6 auto… 4        15    19 r     suv  
#> 3 jeep         grand cher…   4    1999     6 auto… 4        15    20 r     suv  
#> 4 jeep         grand cher…   4.7  1999     8 auto… 4        14    17 r     suv  
#> 5 jeep         grand cher…   4.7  2008     8 auto… 4         9    12 e     suv  
#> 6 jeep         grand cher…   4.7  2008     8 auto… 4        14    19 r     suv  
#> 7 jeep         grand cher…   5.7  2008     8 auto… 4        13    18 r     suv  
#> 8 jeep         grand cher…   6.1  2008     8 auto… 4        11    14 p     suv  
#> 
#> $`land rover`
#> # A tibble: 4 × 11
#>   manufacturer model       displ  year   cyl trans drv     cty   hwy fl    class
#>   <chr>        <chr>       <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#> 1 land rover   range rover   4    1999     8 auto… 4        11    15 p     suv  
#> 2 land rover   range rover   4.2  2008     8 auto… 4        12    18 r     suv  
#> 3 land rover   range rover   4.4  2008     8 auto… 4        12    18 r     suv  
#> 4 land rover   range rover   4.6  1999     8 auto… 4        11    15 p     suv  
#> 
#> $lincoln
#> # A tibble: 3 × 11
#>   manufacturer model       displ  year   cyl trans drv     cty   hwy fl    class
#>   <chr>        <chr>       <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#> 1 lincoln      navigator …   5.4  1999     8 auto… r        11    17 r     suv  
#> 2 lincoln      navigator …   5.4  1999     8 auto… r        11    16 p     suv  
#> 3 lincoln      navigator …   5.4  2008     8 auto… r        12    18 r     suv  
#> 
#> $mercury
#> # A tibble: 4 × 11
#>   manufacturer model       displ  year   cyl trans drv     cty   hwy fl    class
#>   <chr>        <chr>       <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#> 1 mercury      mountainee…   4    1999     6 auto… 4        14    17 r     suv  
#> 2 mercury      mountainee…   4    2008     6 auto… 4        13    19 r     suv  
#> 3 mercury      mountainee…   4.6  2008     8 auto… 4        13    19 r     suv  
#> 4 mercury      mountainee…   5    1999     8 auto… 4        13    17 r     suv  
#> 
#> $nissan
#> # A tibble: 13 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 nissan       altima       2.4  1999     4 manu… f        21    29 r     comp…
#>  2 nissan       altima       2.4  1999     4 auto… f        19    27 r     comp…
#>  3 nissan       altima       2.5  2008     4 auto… f        23    31 r     mids…
#>  4 nissan       altima       2.5  2008     4 manu… f        23    32 r     mids…
#>  5 nissan       altima       3.5  2008     6 manu… f        19    27 p     mids…
#>  6 nissan       altima       3.5  2008     6 auto… f        19    26 p     mids…
#>  7 nissan       maxima       3    1999     6 auto… f        18    26 r     mids…
#>  8 nissan       maxima       3    1999     6 manu… f        19    25 r     mids…
#>  9 nissan       maxima       3.5  2008     6 auto… f        19    25 p     mids…
#> 10 nissan       pathfinde…   3.3  1999     6 auto… 4        14    17 r     suv  
#> 11 nissan       pathfinde…   3.3  1999     6 manu… 4        15    17 r     suv  
#> 12 nissan       pathfinde…   4    2008     6 auto… 4        14    20 p     suv  
#> 13 nissan       pathfinde…   5.6  2008     8 auto… 4        12    18 p     suv  
#> 
#> $pontiac
#> # A tibble: 5 × 11
#>   manufacturer model      displ  year   cyl trans  drv     cty   hwy fl    class
#>   <chr>        <chr>      <dbl> <int> <int> <chr>  <chr> <int> <int> <chr> <chr>
#> 1 pontiac      grand prix   3.1  1999     6 auto(… f        18    26 r     mids…
#> 2 pontiac      grand prix   3.8  1999     6 auto(… f        16    26 p     mids…
#> 3 pontiac      grand prix   3.8  1999     6 auto(… f        17    27 r     mids…
#> 4 pontiac      grand prix   3.8  2008     6 auto(… f        18    28 r     mids…
#> 5 pontiac      grand prix   5.3  2008     8 auto(… f        16    25 p     mids…
#> 
#> $subaru
#> # A tibble: 14 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 subaru       forester …   2.5  1999     4 manu… 4        18    25 r     suv  
#>  2 subaru       forester …   2.5  1999     4 auto… 4        18    24 r     suv  
#>  3 subaru       forester …   2.5  2008     4 manu… 4        20    27 r     suv  
#>  4 subaru       forester …   2.5  2008     4 manu… 4        19    25 p     suv  
#>  5 subaru       forester …   2.5  2008     4 auto… 4        20    26 r     suv  
#>  6 subaru       forester …   2.5  2008     4 auto… 4        18    23 p     suv  
#>  7 subaru       impreza a…   2.2  1999     4 auto… 4        21    26 r     subc…
#>  8 subaru       impreza a…   2.2  1999     4 manu… 4        19    26 r     subc…
#>  9 subaru       impreza a…   2.5  1999     4 manu… 4        19    26 r     subc…
#> 10 subaru       impreza a…   2.5  1999     4 auto… 4        19    26 r     subc…
#> 11 subaru       impreza a…   2.5  2008     4 auto… 4        20    25 p     comp…
#> 12 subaru       impreza a…   2.5  2008     4 auto… 4        20    27 r     comp…
#> 13 subaru       impreza a…   2.5  2008     4 manu… 4        19    25 p     comp…
#> 14 subaru       impreza a…   2.5  2008     4 manu… 4        20    27 r     comp…
#> 
#> $toyota
#> # A tibble: 34 × 11
#>    manufacturer model      displ  year   cyl trans drv     cty   hwy fl    class
#>    <chr>        <chr>      <dbl> <int> <int> <chr> <chr> <int> <int> <chr> <chr>
#>  1 toyota       4runner 4…   2.7  1999     4 manu… 4        15    20 r     suv  
#>  2 toyota       4runner 4…   2.7  1999     4 auto… 4        16    20 r     suv  
#>  3 toyota       4runner 4…   3.4  1999     6 auto… 4        15    19 r     suv  
#>  4 toyota       4runner 4…   3.4  1999     6 manu… 4        15    17 r     suv  
#>  5 toyota       4runner 4…   4    2008     6 auto… 4        16    20 r     suv  
#>  6 toyota       4runner 4…   4.7  2008     8 auto… 4        14    17 r     suv  
#>  7 toyota       camry        2.2  1999     4 manu… f        21    29 r     mids…
#>  8 toyota       camry        2.2  1999     4 auto… f        21    27 r     mids…
#>  9 toyota       camry        2.4  2008     4 manu… f        21    31 r     mids…
#> 10 toyota       camry        2.4  2008     4 auto… f        21    31 r     mids…
#> # ℹ 24 more rows
#> 
#> $volkswagen
#> # A tibble: 27 × 11
#>    manufacturer model displ  year   cyl trans      drv     cty   hwy fl    class
#>    <chr>        <chr> <dbl> <int> <int> <chr>      <chr> <int> <int> <chr> <chr>
#>  1 volkswagen   gti     2    1999     4 manual(m5) f        21    29 r     comp…
#>  2 volkswagen   gti     2    1999     4 auto(l4)   f        19    26 r     comp…
#>  3 volkswagen   gti     2    2008     4 manual(m6) f        21    29 p     comp…
#>  4 volkswagen   gti     2    2008     4 auto(s6)   f        22    29 p     comp…
#>  5 volkswagen   gti     2.8  1999     6 manual(m5) f        17    24 r     comp…
#>  6 volkswagen   jetta   1.9  1999     4 manual(m5) f        33    44 d     comp…
#>  7 volkswagen   jetta   2    1999     4 manual(m5) f        21    29 r     comp…
#>  8 volkswagen   jetta   2    1999     4 auto(l4)   f        19    26 r     comp…
#>  9 volkswagen   jetta   2    2008     4 auto(s6)   f        22    29 p     comp…
#> 10 volkswagen   jetta   2    2008     4 manual(m6) f        21    29 p     comp…
#> # ℹ 17 more rows
#> 

# Multiple grouping columns:
parts <- split_by(ggplot2::mpg, manufacturer, cyl)
names(parts)
#>  [1] "audi_4"       "audi_6"       "audi_8"       "chevrolet_4"  "chevrolet_6" 
#>  [6] "chevrolet_8"  "dodge_4"      "dodge_6"      "dodge_8"      "ford_6"      
#> [11] "ford_8"       "honda_4"      "hyundai_4"    "hyundai_6"    "jeep_6"      
#> [16] "jeep_8"       "land rover_8" "lincoln_8"    "mercury_6"    "mercury_8"   
#> [21] "nissan_4"     "nissan_6"     "nissan_8"     "pontiac_6"    "pontiac_8"   
#> [26] "subaru_4"     "toyota_4"     "toyota_6"     "toyota_8"     "volkswagen_4"
#> [31] "volkswagen_5" "volkswagen_6"